\name{ngram-package}
\alias{ngram-package}
\docType{package}
\title{ngram: An n-gram Babbler}
\description{
  This package offers utilities for creating, displaying,
  and "babbling" n-grams.  The babbler is a simple Markov
  process.
}
\details{
\tabular{ll}{
  Package: \tab ngram\cr
  Type: \tab Package\cr
  License: \tab 2-clause BSD\cr
}
}
\references{
  ref
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}, Christian
  Heckendorf.
}
\keyword{Package}
