\name{Utilities}

\alias{concat}
\alias{wordcount}

\title{n-gram Utilities}

\description{
  Some utility methods.
}
\usage{
  concat(str)
  wordcount(str)
}
\arguments{
  \item{str}{A string or vector of strings.}
}
\details{
  deets.
}
\value{
  \code{concat()} returns 
}
\seealso{
  \code{\link{ngram-class}}
}
\examples{
\dontrun{
library(ngram)

str <- "a b c"
concat(str)
wordcount(str)
}
}
