% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/charsplitter.r
\name{charsplitter}
\alias{charsplitter}
\title{Character Splitter}
\usage{
charsplitter(string, spacechar = "_")
}
\arguments{
\item{string}{An input string.}

\item{spacechar}{The character to represent a space.}
}
\value{
A string.
}
\description{
A utility function for use with n-gram modeling.  This function
splits a string without spaces into one useable with \code{ngram()}.
}
\examples{
x <- "abacabb"
charsplitter(x)

y <- "split at every character, preserve spaces"
charsplitter(y)
}
\keyword{Preprocessing}

