% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{Concatenate}
\alias{Concatenate}
\alias{concat}
\title{Concatenate}
\usage{
concat(..., collapse = "", rm.spaces = FALSE)
}
\arguments{
\item{...}{Input text(s).}

\item{collapse}{A character to separate the input strings if a vector of
strings is supplied; otherwisethis does nothing.}

\item{rm.spaces}{logical; determines if spaces should be removed from the
final string.}
}
\value{
A string.
}
\description{
This is a utility for concatenating strings together.  This is
handy because if you want to generate the n-grams for several different
texts, you must first put them into a single string.
}
\examples{
\dontrun{
library(ngram)

words <- c("a", "b", "c")
wordcount(words)
concat(words)
}
}
\seealso{
\code{\link{Preprocess}}
}

