\name{process}

\docType{methods}

\alias{ngram-method}
\alias{ngram,character-method}
\alias{ngram}

\title{ngram}

\description{
  processing
}
\section{Methods}{
\describe{
\item{\code{signature(str = "character")}}{}
}
}
\usage{
\S4method{ngram}{character}(str, n=2)
}
\arguments{
  \item{str}{The input string.}
  \item{n}{The 'n' as in 'n-gram'.}
}
\details{
  deets.
}
\value{
  something
}
\seealso{
  \code{\link{ngram}}
}
\keyword{Data Generation}


