\name{ngram Getters}

\docType{methods}

\alias{get.ngrams-method}
\alias{get.ngrams,ngram-method}
\alias{get.ngrams}

\alias{get.string-method}
\alias{get.string,ngram-method}
\alias{get.string}

\alias{get.nextwords-method}
\alias{get.nextwords,ngram-method}
\alias{get.nextwords}

\title{ngram Getters}

\description{
  Getters
}
\section{Methods}{
\describe{
\item{\code{signature(ng = "ngram")}}{}
}
}
\usage{
\S4method{get.ngrams}{ngram}(ng)
\S4method{get.string}{ngram}(ng)
\S4method{get.nextwords}{ngram}(ng)
}
\arguments{
  \item{ng}{An ngram object.}
}
\details{
  deets.
}
\value{
  something
}
\seealso{
  \code{\link{ngram}}
}
\keyword{Data Generation}


