\name{ngram Babbler}

\docType{methods}

\alias{babble-method}
\alias{babble,ngram-method}
\alias{babble}

\title{ngram Getters}

\description{
  babbler
}
\section{Methods}{
\describe{
\item{\code{signature(ng = "ngram")}}{}
}
}
\usage{
\S4method{babble}{ngram}(ng, genlen=150, seed=getseed()))
}
\arguments{
  \item{ng}{An ngram object.}
  \item{genlen}{.}
  \item{seed}{.}
}
\details{
  deets.
}
\value{
  something
}
\seealso{
  \code{\link{ngram}}
}
\keyword{Data Generation}


