% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phrasetable.r
\name{get.phrasetable}
\alias{get.phrasetable}
\title{Get Phrasetable}
\usage{
get.phrasetable(ng)
}
\arguments{
\item{ng}{An ngram object.}
}
\description{
Get a table
}
\examples{
\dontrun{
library(ngram)

str <- "A B A C A B B"
ng <- ngram(str)
get.phrasetable(ng)
}
}
\seealso{
\code{\link{ngram-class}}
}

