% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseed.r
\name{getseed}
\alias{getseed}
\title{getseed}
\usage{
getseed()
}
\description{
A seed generator for use with the ngram package.
}
\details{
Uses a 96-bit hash of the current process id, time, and a random
uniform value from R's random generator.
}
\seealso{
\code{\link{babble}}
}
\keyword{Utility}
