% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcorpus.r
\name{rcorpus}
\alias{rcorpus}
\title{Random Corpus}
\usage{
rcorpus(nwords, minwordlen = 1, maxwordlen = 6)
}
\arguments{
\item{nwords}{Number of words to generate.}

\item{minwordlen,maxwordlen}{The min/max length of words in the generated corpus.}
}
\value{
A string.
}
\description{
Generate a corpus of random "words".
}
\examples{
rcorpus(10)
}

