% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.summary.r
\name{string.summary}
\alias{string.summary}
\title{Text Summary}
\usage{
string.summary(string, wordlen_max = 10, senlen_max = 10, syllen_max = 10)
}
\arguments{
\item{string}{An input string.}

\item{wordlen_max, senlen_max, syllen_max}{The maximum lengths of words/sentences/syllables to consider.}
}
\value{
A list of class \code{string_summary}.
}
\description{
Text Summary
}
\examples{
x <- "a b a c a b b"

string.summary(x)

}
\keyword{Summarize}

