\name{ngram-package}
\alias{ngram-package}
\docType{package}
\title{ngram: An n-gram Babbler}
\description{
  This package offers utilities for creating, displaying,
  and "babbling" n-grams.  The babbler is a simple Markov
  process.
}
\details{
  The ngram package is distributed under the permissive 2-clause BSD 
  license.  If you find the code here useful, please let us know
  and/or cite the package, whatever is appropriate.
  
  The package has its own PRNG; we use an implementation of MT1997
  for all non-deterministic choices.
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}, Christian
  Heckendorf.
}
\keyword{Package}
