% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/concat.r
\name{concat}
\alias{concat}
\title{concat}
\usage{
concat(..., collapse = "", rm.space = FALSE)
}
\arguments{
\item{...}{Input text(s).}

\item{collapse}{A character to separate the input strings if a vector of
strings is supplied; otherwisethis does nothing.}

\item{rm.space}{logical; determines if spaces should be removed from the
final string.}
}
\value{
A string.
}
\description{
\code{concat()} is a utility for concatenating strings together.  This is
handy because if you want to generate the n-grams for several different
texts, you must first put them into a single string.
}
\examples{
\dontrun{
library(ngram)

words <- c("a", "b", "c")
wordcount(words)
str <- concat(words)
wordcount(str)
}
}
\seealso{
\code{\link{Preprocess}}
}

