\name{ngram print}

\docType{methods}

\alias{print-method}
\alias{print,ngram-method}
\alias{print}

\alias{show-method}
\alias{show,ngram-method}
\alias{show}

\title{ngram printing}

\description{
  Getters
}
\section{Methods}{
\describe{
\item{\code{signature(ng = "ngram")}}{}
}
}
\usage{
\S4method{print}{ngram}(ng, full=FALSE)
\S4method{show}{ngram}(ng)
}
\arguments{
  \item{ng}{An ngram object.}
  \item{full}{.}
}
\details{
  deets.
}
\value{
  something
}
\seealso{
  \code{\link{ngram}}
}
\keyword{Data Generation}


