% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorpus.r
\name{rcorpus}
\alias{rcorpus}
\title{Random Corpus}
\usage{
rcorpus(nwords = 50, alphabet = letters, minwordlen = 1, maxwordlen = 6)
}
\arguments{
\item{nwords}{Number of words to generate.}

\item{alphabet}{The pool of "letters" that word generation coes from.  By default,
it is the lowercase roman alphabet.}

\item{minwordlen, maxwordlen}{The min/max length of words in the generated corpus.}
}
\value{
A string.
}
\description{
Generate a corpus of random "words".
}
\examples{
rcorpus(10)

}
\keyword{Utility}

