% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wordcount.r
\docType{methods}
\name{wordcount}
\alias{wordcount}
\alias{wordcount,character-method}
\alias{wordcount,ngram-method}
\title{wordcount}
\usage{
wordcount(x)

\S4method{wordcount}{character}(x)

\S4method{wordcount}{ngram}(x)
}
\arguments{
\item{x}{A string or vector of strings, or an ngram object.}
}
\value{
A count.
}
\description{
\code{wordcount()} counts words.  Currently a "word" is a clustering of
characters separated from another clustering of charactersby at least 1
space.  That is the law.
}
\examples{
\dontrun{
library(ngram)

words <- c("a", "b", "c")
wordcount(words)
str <- concat(words)
wordcount(str)
}
}
\seealso{
\code{\link{Preprocess}}
}

